+++ Assets/Scripts/ScriptableObjects/Upgrades/PassiveUpgrade.cs
@@ -5,7 +5,7 @@ namespace ldjam_58
     [CreateAssetMenu(fileName = "PassiveUpgrade", menuName = "SO/Upgrades/PassiveUpgrade")]
     public class PassiveUpgrade : Upgrade
     {
-        [SerializeField] private float passiveModifier;
+        [SerializeField] private uint passiveModifier;
         [SerializeField] private GameManagerChannel gameManagerChannel;
 
         private void OnEnable()
